COUNTED LOOPS BABY!
Add counted ?DO..LOOP loops. No index tracking yet.
See here for an explanation:
https://www.complang.tuwien.ac.at/forth/gforth/Docs-html/Counted-Loops.html

Note that in ANS Forth, ?DO..LOOP is only supposed to work in a
compiled word definition, but for us it's easier to just make it
work anywhere.

I also downcase all code as soon as it enters the interpreter.
This eliminates a potential source of future bugs in methods
such as consume_until(), that previously needed to remember
to deal with case. I didn't pack these chunks in a separate
commit, because lazy ass.